requirements.txt
sqlalchemy
pyYAML
PyQt5
odfpy
pandas
python-magic
openpyxl
pdfplumber
tldextract
fuzzywuzzy
python-Levenshtein
phonenumbers